requirements.txt
Pillow
ffmpeg-python
ffpyplayer
kivy
gphoto2